Refactor worker dashboard and earnings pages; remove old dashboard code, enhance earnings layout, and implement reviews page with submission functionality.
